backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
PyJWT
scikit-learn
pandas
numpy
ollama
gevent
APScheduler
joblib
